Remove ShibbolethRestController and move code into ShibbolethLoginFilter.
