+++ dspace-server-webapp/src/main/java/org/dspace/app/rest/security/ShibbolethLoginFilter.java
@@ -8,14 +8,19 @@
 package org.dspace.app.rest.security;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import javax.servlet.FilterChain;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.dspace.authenticate.ShibAuthentication;
+import org.dspace.core.Utils;
+import org.dspace.services.ConfigurationService;
+import org.dspace.services.factory.DSpaceServicesFactory;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.ProviderNotFoundException;
 import org.springframework.security.core.Authentication;
@@ -24,15 +29,32 @@ import org.springframework.security.core.AuthenticationException;
 /**
  * This class will filter Shibboleth requests to see if the user has been authenticated via Shibboleth.
  * <P>
- * This filter runs before the ShibbolethRestController, in order to verify Shibboleth authentication succeeded,
- * and create the authentication token (JWT).
+ * The overall Shibboleth login process is as follows:
+ *   1. When Shibboleth plugin is enabled, client/UI receives Shibboleth's absolute URL in WWW-Authenticate header.
+ *      See {@link org.dspace.authenticate.ShibAuthentication} loginPageURL() method.
+ *   2. Client sends the user to that URL when they select Shibboleth authentication.
+ *   3. User logs in using Shibboleth
+ *   4. If successful, they are redirected by Shibboleth to the path where this Filter is "listening" (that path
+ *      is passed to Shibboleth as a URL param in step 1)
+ *   5. This filter then intercepts the request in order to check for a valid Shibboleth login (see
+ *      ShibAuthentication.authenticate()) and stores that user info in a JWT. It also saves that JWT in a *temporary*
+ *      authentication cookie.
+ *   6. This filter then looks for a "redirectUrl" param (also a part of the original URL from step 1), and redirects
+ *      the user to that location (after verifying it's a trusted URL). Usually this is a redirect back to the
+ *      Client/UI page where the User started.
+ *   7. At that point, the client reads the JWT from the Cookie, and sends it back in a request to /api/authn/login,
+ *      which triggers the server-side to destroy the Cookie and move the JWT into a Header
+ * <P>
+ * This Shibboleth Authentication process is tested in AuthenticationRestControllerIT.
  *
  * @author Giuseppe Digilio (giuseppe dot digilio at 4science dot it)
- * @see org.dspace.app.rest.ShibbolethRestController
+ * @author Tim Donohue
  * @see org.dspace.authenticate.ShibAuthentication
  */
 public class ShibbolethLoginFilter extends StatelessLoginFilter {
     private static final Logger log = LogManager.getLogger(ShibbolethLoginFilter.class);
+
+    private ConfigurationService configurationService = DSpaceServicesFactory.getInstance().getConfigurationService();
 
     public ShibbolethLoginFilter(String url, AuthenticationManager authenticationManager,
                                  RestAuthenticationService restAuthenticationService) {
@@ -49,9 +71,9 @@ public class ShibbolethLoginFilter extends StatelessLoginFilter {
         }
 
         // In the case of Shibboleth, this method does NOT actually authenticate us. The authentication
-        // has already happened in Shibboleth & we are just intercepting the return request in order to check
-        // for a valid Shibboleth login (using ShibAuthentication.authenticate()) & save current user to Context
-        // See org.dspace.app.rest.ShibbolethRestController JavaDocs for an outline of the entire Shib login process.
+        // has already happened in Shibboleth. So, this call to "authenticate()" is just triggering
+        // ShibAuthentication.authenticate() to check for a valid Shibboleth login, and if found, the current user
+        // is considered authenticated via Shibboleth.
         // NOTE: because this authentication is implicit, we pass in an empty DSpaceAuthentication
         return authenticationManager.authenticate(new DSpaceAuthentication());
     }
@@ -65,12 +87,51 @@ public class ShibbolethLoginFilter extends StatelessLoginFilter {
 
         DSpaceAuthentication dSpaceAuthentication = (DSpaceAuthentication) auth;
         // OVERRIDE DEFAULT behavior of StatelessLoginFilter to return a temporary authentication cookie containing
-        // the Auth Token (JWT). This Cookie is required because ShibbolethRestController *redirects* the user
-        // back to the client/UI after a successful Shibboleth login. Headers cannot be sent via a redirect, so a Cookie
-        // must be sent to provide the auth token to the client. On the next request from the client, the cookie is
-        // read and destroyed & the Auth token is only used in the Header from that point forward.
+        // the Auth Token (JWT). This Cookie is required because we *redirect* the user back to the client/UI after
+        // a successful Shibboleth login. Headers cannot be sent via a redirect, so a Cookie must be sent to provide
+        // the auth token to the client. On the next request from the client, the cookie is read and destroyed & the
+        // Auth token is only used in the Header from that point forward.
         restAuthenticationService.addAuthenticationDataForUser(req, res, dSpaceAuthentication, true);
-        chain.doFilter(req, res);
+
+        // redirect user after completing Shibboleth authentication, sending along the temporary auth cookie
+        redirectAfterSuccess(req, res);
     }
 
+
+    /**
+     * After successful login, redirect to the DSpace URL specified by this Shibboleth request (in the "redirectUrl"
+     * request parameter). If that 'redirectUrl' is not valid or trusted for this DSpace site, then return a 400 error.
+     * @param request
+     * @param response
+     * @throws IOException
+     */
+    private void redirectAfterSuccess(HttpServletRequest request, HttpServletResponse response) throws IOException {
+        // Get redirect URL from request parameter
+        String redirectUrl = request.getParameter("redirectUrl");
+
+        // If redirectUrl unspecified, default to the configured UI
+        if (StringUtils.isEmpty(redirectUrl)) {
+            redirectUrl = configurationService.getProperty("dspace.ui.url");
+        }
+
+        // Validate that the redirectURL matches either the server or UI hostname. It *cannot* be an arbitrary URL.
+        String redirectHostName = Utils.getHostName(redirectUrl);
+        String serverHostName = Utils.getHostName(configurationService.getProperty("dspace.server.url"));
+        ArrayList<String> allowedHostNames = new ArrayList<>();
+        allowedHostNames.add(serverHostName);
+        String[] allowedUrls = configurationService.getArrayProperty("rest.cors.allowed-origins");
+        for (String url : allowedUrls) {
+            allowedHostNames.add(Utils.getHostName(url));
+        }
+
+        if (StringUtils.equalsAnyIgnoreCase(redirectHostName, allowedHostNames.toArray(new String[0]))) {
+            log.debug("Shibboleth redirecting to " + redirectUrl);
+            response.sendRedirect(redirectUrl);
+        } else {
+            log.error("Invalid Shibboleth redirectURL=" + redirectUrl +
+                          ". URL doesn't match hostname of server or UI!");
+            response.sendError(HttpServletResponse.SC_BAD_REQUEST,
+                               "Invalid redirectURL! Must match server or ui hostname.");
+        }
+    }
 }
